Modified PRI fit which adjusts lobe intensities

+++ misc/diskCleanup.py
@@ -2,9 +2,9 @@
 import sys
 import os
 
-micrDataDir = '/mnt/MicrData/'
+micrDataDir = '/smb/MicrData/'
 lDataDir = len(micrDataDir)
-nasDir = '/mnt/NasData/'
+nasDir = '/smb/NasData/'
 #backupDir = '/media/disk/LMNAS1/'
 
 def deleteFiles(directory):
@@ -15,11 +15,11 @@ def deleteFiles(directory):
             #print filename
 
             nFilename = nasDir + filename[lDataDir:]
-            #bFilename = backupDir + filename[lDataDir:]
+            bFilename = backupDir + filename[lDataDir:]
 
-            #print nFilename, bFilename
+            print nFilename, bFilename
 
-            if os.path.exists(nFilename): #and os.path.exists(bFilename) and os.path.getsize(filename) == os.path.getsize(nFilename):
+            if os.path.exists(nFilename) and os.path.getsize(nFilename) == os.path.getsize(filename): #and os.path.exists(bFilename) and os.path.getsize(filename) == os.path.getsize(nFilename):
                 print 'Deleting %s' % filename
                 os.remove(filename)
             else:
@@ -28,7 +28,7 @@ def deleteFiles(directory):
         for dir in dirs:
             dirname = os.path.join(path, dir)
 
-            if len(os.listdir(dirname)) == 0:
+            if len(os.listdir(dirname)) == 0 and not 'System Volume Information' in dirname:
                 print 'Removing %s' % dirname
                 os.rmdir(dirname)
